requirements.txt
chris_plugin==0.4.0
pydicom
numpy
pillow
# for bug fix on transfer syntax errors
pylibjpeg
pylibjpeg-libjpeg
python-gdcm
